create three component CardWithHeading,CardWithRating and IconHorizontalList and add bottom tab bar
