+++ package.json
@@ -16,6 +16,7 @@
     "@react-native-firebase/auth": "^11.2.0",
     "@react-native-firebase/messaging": "^11.2.0",
     "@react-native-google-signin/google-signin": "^6.0.0",
+    "@react-navigation/bottom-tabs": "^5.11.9",
     "@react-navigation/drawer": "^5.12.5",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.4",
